requirement.txt
accelerate==0.13.2
certifi==2022.9.24
charset-normalizer==2.1.1
diffusers==0.7.1
filelock==3.8.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
psutil==5.9.3
pyparsing==3.0.9
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
